requirements.txt
django==3.0.5
djangorestframework==3.10.0
pytest==5.3.5
pytest-django==3.8.0
psycopg2==2.8.5
django-dotenv==1.4.2
